+++ index.js
@@ -24,20 +24,28 @@ module.exports = function(css, options){
    */
 
   function position() {
-    var start = { line: lineno, column: column };
     if (!options.position) return positionNoop;
 
-    return function(node){
-      node.position = {
-        start: start,
-        end: { line: lineno, column: column },
-        source: options.source
-      };
+    var start = { line: lineno, column: column };
 
+    return function(node){
+      node.position = new Position(start);
       whitespace();
       return node;
     }
   }
+
+  function Position(start) {
+    this.start = start;
+    this.end = { line: lineno, column: column };
+    this.filename = options.filename;
+  }
+
+  /**
+   * Non-enumerable source string.
+   */
+
+  Position.prototype.source = css;
 
   /**
    * Return `node`.
@@ -52,12 +60,9 @@ module.exports = function(css, options){
    * Error `msg`.
    */
 
-  function error(msg) {
+  function error(msg, start) {
     var err = new Error(msg + ' near line ' + lineno + ':' + column);
-    err.filename = options.source;
-    err.line = lineno;
-    err.column = column;
-    err.source = css;
+    err.position = new Position(start);
     throw err;
   }
 
@@ -169,7 +174,7 @@ module.exports = function(css, options){
   function selector() {
     var m = match(/^([^{]+)/);
     if (!m) return;
-    /* @fix Remove all comments from selectors 
+    /* @fix Remove all comments from selectors
      * http://ostermiller.org/findcomment.html */
     return trim(m[0]).replace(/\/\*([^*]|[\r\n]|(\*+([^*/]|[\r\n])))*\*\/+/g, '').split(/\s*,\s*/);
   }